Move ReportingListSummary into Reporting feature components
