+++ frontend/src/features/Reporting/components/ReportingListSummary/InfractionSuspicionSummary.tsx
